+++ payments/payments.go
@@ -24,14 +24,14 @@ type Provider interface {
 }
 
 // Charger wraps the Charge method which creates new payments with the provider.
-type Charger func(amount uint64, currency string, order *models.Order, invoiceNumber int64) (string, error)
+type Charger func(amount float64, currency string, order *models.Order, invoiceNumber int64) (string, error)
 
 // Refunder wraps the Refund method which refunds payments with the provider.
-type Refunder func(transactionID string, amount uint64, currency string) (string, error)
+type Refunder func(transactionID string, amount float64, currency string) (string, error)
 
 // Preauthorizer wraps the Preauthorize method which pre-authorizes a payment
 // with the provider.
-type Preauthorizer func(amount uint64, currency string, description string) (*PreauthorizationResult, error)
+type Preauthorizer func(amount float64, currency string, description string) (*PreauthorizationResult, error)
 
 // PreauthorizationResult contains the data returned from a Preauthorization.
 type PreauthorizationResult struct {
